pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rhetor"
version = "1.0.0-dev"
description = "The Master Orator for Your Documents — a Windows desktop document reader with TTS"
requires-python = ">=3.10,<3.13"
license = {text = "Proprietary"}
authors = [
    {name = "AJ Geddes", email = "aj@highvelocitysolutions.com"},
]

dependencies = [
    "pymupdf>=1.24.0",
    "python-docx>=1.1.0",
    "charset-normalizer>=3.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4.0",
    "mypy>=1.10",
]

[tool.setuptools.packages.find]
include = ["core*", "tts*", "audio*", "ui*"]

[tool.ruff]
target-version = "py310"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
